refactor: casts max log file size to integer in configuration Ensures consistent handling of log file size in the application.

+++ src/config/filament-log-viewer.php
@@ -11,7 +11,7 @@ return [
     | The maximum size (in kilobytes) for a single log file to be read.
     | Files larger than this will be skipped to prevent memory exhaustion.
     */
-    'max_log_file_size' => env('LOG_MAX_SIZE_KB', 2048),
+    'max_log_file_size' => (int) env('LOG_MAX_SIZE_KB', 2048),
 
     /*
     |--------------------------------------------------------------------------
